% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{add.heatmap}
\alias{add.heatmap}
\title{add.heatmap
Description: Add oligprofile heatmap into output directory}
\usage{
add.heatmap(dat, output.dir, output.file = NULL, taxonomy, ppcm = 150,
  hclust.method = "complete", palette = "white/black", level = "L1",
  metric = "pearson", figureratio = 10, fontsize = 40,
  tree.display = TRUE)
}
\description{
Arguments:
  @param dat oligoprofile data in original (non-log) domain
  @param output.dir output data directory
  @param output.file output file name
  @param taxonomy oligo-phylotype mappings
  @param ppcm figure size
  @param hclust.method hierarchical clustering method
  @param palette color palette ("white/black" / "white/blue" / "black/yellow/white")
  @param level taxonomic level to show
  @param metric clustering metric
  @param figureratio figure ratio
  @param fontsize font size
  @param tree.display tree.display
}
\details{
Returns:
  @return Plotting parameters
}
\examples{
# data(peerj32); hc <- add.heatmap(peerj32$microbes[, 1:4])
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

