% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bootstrap.microbes}
\alias{bootstrap.microbes}
\title{bootstrap.microbes}
\usage{
bootstrap.microbes(D, Nsample = NULL, minPrev = 2, Nboot = 1000,
  I.thr = 1.8, ncore = 1)
}
\arguments{
\item{D}{data (phylotypes x samples)}

\item{Nsample}{bootstrap sample size, default is the same size as data}

\item{minPrev}{Lower limit for number of samples where microbe needs to exceed the intensity threshold for a 'present' call.}

\item{Nboot}{bootstrap sample size}

\item{I.thr}{Lower limit for intensity threshold}

\item{ncore}{number of nodes for parallelization - default 1

Returns:}
}
\value{
data frame with microbes and their frequency of presence in the core
}
\description{
Description: bootstrap.microbes
}
\details{
Arguments:
}
\examples{
\dontrun{data(peerj32);
	     bs <- bootstrap.microbes(t(peerj32$microbes), Nboot = 5)}
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

