% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{calculate.hclust}
\alias{calculate.hclust}
\title{calculate.hclust}
\usage{
calculate.hclust(dat, method = "complete", metric = "pearson")
}
\description{
Calculate hierarchical clustering for standard selections in 
profiling script
}
\details{
@param dat data matrix (use log10 with pearson!)
  @param method hierarchical clustering method (see ?hclust)
  @param metric clustering metrics (spearman / pearson / euclidean)

@return hclust object for log10 and for absolute scale data
}
\examples{
\dontrun{  
  data(peerj32)
  dat <- peerj32$microbes
  hc <- calculate.hclust(dat, 'complete', 'pearson')
}
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

