% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{choose.projects}
\alias{choose.projects}
\title{Description: Select projects to analyze}
\usage{
choose.projects(con, multi = TRUE, condition = NULL)
}
\description{
Arguments:
  @param con valid MySQL connection
  @param multi enable selection of multiple options
  @param condition TBA
Returns:
  @return vector of project names
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

