\name{chooseDir}
\alias{chooseDir}
\title{Choosing (and creating) a directory}
\usage{
  chooseDir(...)
}
\arguments{
  \item{...}{parameters to pass}
}
\value{
  choice
}
\description{
  Choosing (and creating) a directory
}
\examples{
# mydir <- chooseDir()
}
\author{
<<<<<<< HEAD
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{microbiome-admin@googlegroups.com} importFrom
  svDialogs guiDlgDir importFrom svDialogs dir.create
=======
Douwe Molenaar. Maintainer: Leo Lahti
\email{microbiome-admin@googlegroups.com}
>>>>>>> 22cc689b4272a98f850a07424aff96d103c59b88
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

