% Generated by roxygen2 (4.0.0): do not edit by hand
\name{count}
\alias{count}
\title{Description: count}
\usage{
count(d)
}
\arguments{
\item{d}{TBA

Returns:}
}
\value{
list
}
\description{
For cross-hyb control
}
\details{
Arguments:
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

