% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.chip.R
\name{detect.chip}
\alias{detect.chip}
\title{Description: Detect the chip type (H/M/PITChip) from database name}
\usage{
detect.chip(dbname)
}
\description{
Arguments:
  @param dbname MySQL database name
}
\details{
Returns:
  @return chip name
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

