% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{estimate.min.threshold}
\alias{estimate.min.threshold}
\title{Description: determine threshold for bg correction}
\usage{
estimate.min.threshold(dat)
}
\description{
Arguments:
  @param dat data matrix (in approximately normal scale ie. logged)
}
\details{
Returns:
  @return threshold value
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

