% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{expandCondition}
\alias{expandCondition}
\title{Expands a list of "field", "value" pair lists into an SQL condition}
\usage{
expandCondition(condition)
}
\description{
Arguments:
 @param condition condition
}
\details{
Returns:
 @return TBA
}
\examples{
# TBA
}
\author{
Douwe Molenaar. Maintainer: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

