% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{expandElement}
\alias{expandElement}
\title{Expands one element (one "field", "value" pair list) from a list of "field", "value" pair lists}
\usage{
expandElement(elm)
}
\arguments{
\item{elm}{TBA}
}
\value{
TBA
}
\description{
Expands one element (one "field", "value" pair list) from a list of "field", "value" pair lists
}
\author{
Douwe Molenaar. Maintainer: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

