% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.sample.info.R
\name{fetch.sample.info}
\alias{fetch.sample.info}
\title{fetch.sample.info}
\usage{
fetch.sample.info(allowed.projects, chiptype = NULL, dbuser, dbpwd, dbname,
  selected.samples = NULL, host = NULL, port = NULL)
}
\description{
Description: Fetch sample information from HITChip atlas
}
\details{
Arguments:
  @param allowed.projects list projects for which to fetch the data
  @param chiptype chiptype (eg. new.chip)
  @param dbuser MySQL user
  @param dbpwd MySQL password
  @param dbname MySqL database name
  @param selected.samples Sample to investigate. By default all.
  @param host host; needed with FTP connections
  @param port port; needed with FTP connections
  @return project.info data.frame
}
\examples{
# info <- fetch.sample.info(allowed.projects, dbuser, dbpwd, 
         # dbname, selected.samples = NULL, host = NULL, port = NULL)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

