% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{FetchData}
\alias{FetchData}
\title{Description: Fetch data from the database}
\usage{
FetchData(params, con, scriptVersion, save.data, scaling, cmetrics)
}
\description{
Arguments:
  @param params params 
  @param con con
  @param scriptVersion scriptVersion
  @param save.data save.data
  @param scaling scaling
  @param cmetrics cmetrics
}
\details{
Returns:
  @return data
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

