% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get.doligo2}
\alias{get.doligo2}
\title{Description: Calculate d.oligo2}
\usage{
get.doligo2(featuretab, d.scaled, oligo.ids)
}
\description{
Arguments:
  @param featuretab featuretab
  @param d.scaled d.scaled
  @param oligo.ids oligo.ids
Returns:
  @return d.oligo2
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

