% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.probedata.R
\name{get.probedata}
\alias{get.probedata}
\title{Description: Get probedata}
\usage{
get.probedata(hybridization.ids, rmoligos, dbuser, dbpwd, dbname, host = NULL,
  port = NULL)
}
\description{
Arguments:
  @param hybridization.ids Specify the hybridizations to retrieve
  @param rmoligos oligos to exclude
  @param dbuser MySQL user
  @param dbpwd  MySQL password
  @param dbname MySqL database name
  @param host host; needed with FTP connections
  @param port port; needed with FTP connections
Returns:                                        
  @return list with data (features x hybridizations matrix) and info (features x info) fields
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

