% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get.sampleid}
\alias{get.sampleid}
\title{Description: Pick sampleIDs from d.oligo column names}
\usage{
get.sampleid(d.oligo)
}
\description{
Arguments:
  @param d.oligo d.oligo matrix
Returns:
  @return sampleID vector
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

