% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotPhylochip.R
\name{list.color.scales}
\alias{list.color.scales}
\title{Shade et al. (2014). Conditionally Rare Taxa Disproportionately Contribute to Temporal Changes in Microbial Diversity. mBio 5(4):e01371-14. doi: 10.1128/mBio.01371-14}
\usage{
list.color.scales()
}
\value{
list of color scales
}
\description{
Ellison AM (1987). Effect of seed dimorphism on the density-dependent dynamics of experimental populations of Atriplex triangularis (Chenopodiaceae). Am. J. Bot. 74:1280–1288. doi:10.2307/2444163.
}
\details{
List color scales
}
\examples{
list.color.scales()
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

