% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{list.mysql.projects}
\alias{list.mysql.projects}
\title{list.mysql.projects}
\usage{
list.mysql.projects(dbuser, dbpwd, dbname, host = NULL, port = NULL)
}
\description{
Description: List projects in MySQL database
}
\details{
Arguments:
  @param dbuser MySQL user
  @param dbpwd  MySQL password
  @param dbname MySqL database name
  @param host host; needed with FTP connections
  @param port port; needed with FTP connections
Returns:
  @return project names vector
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

