% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{mysql.format}
\alias{mysql.format}
\title{Description: Format string vector to mysql query format}
\usage{
mysql.format(s)
}
\arguments{
\item{s}{string vector

Returns:
  @return mysql query version}
}
\description{
Arguments:
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
