% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune16S.R
\name{n.phylotypes.per.oligo}
\alias{n.phylotypes.per.oligo}
\title{Description: Check number of matching phylotypes for each probe}
\usage{
n.phylotypes.per.oligo(taxonomy, level)
}
\arguments{
\item{taxonomy}{oligo - phylotype matching data.frame}

\item{level}{phylotype level

Returns:
  @return number of matching phylotypes for each probe}
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

