% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{oligo.bg.correction}
\alias{oligo.bg.correction}
\title{Description: Calculate species summaries and possibly update d.oligo2}
\usage{
oligo.bg.correction(d.oligo2, bgc.method)

oligo.bg.correction(d.oligo2, bgc.method)
}
\description{
Arguments:
  @param d.oligo2 d.oligo2
  @param bgc.method background correction method
Returns:
  @return Background-corrected data matrix

Arguments:
  @param d.oligo2 d.oligo2
  @param bgc.method background correction method
Returns:
  @return Background-corrected data matrix
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}

Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")

See citation("microbiome")
}
\keyword{internal}
\keyword{utilities}

