% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{phylotype.rm.list}
\alias{phylotype.rm.list}
\title{Description: Default list of removed phylotypes and oligos}
\usage{
phylotype.rm.list(chip)
}
\description{
Arguments:
 @param chip Chip name (HIT/MIT/PIT/Chick)Chip
Returns:
  @return List of removed oligos and phylotypes
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

