% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPhylochip.R
\name{PlotPhylochipHeatmap}
\alias{PlotPhylochipHeatmap}
\title{PlotPhylochipHeatmap}
\usage{
PlotPhylochipHeatmap(data, taxonomy, metric = "pearson", level = "L1",
  tree.display = TRUE, palette = "white/black", fontsize = 40,
  figureratio = 10, hclust.method = "complete")
}
\description{
Description: Plots heatmap of the oligo profiles together with phylotype grouping and sample clustering
}
\details{
Arguments:
  @param data oligoprofile data in original (non-log) domain
  @param taxonomy oligo-phylotype mappings
  @param metric clustering metric
  @param level taxonomic level to show (L0 / L1 / L2 / species)
  @param tree.display tree.display
  @param palette color palette ("white/black" / "white/blue" / "black/yellow/white")
  @param fontsize font size
  @param figureratio figure ratio
  @param hclust.method hierarchical clustering method. See help(hclust) for details. To prevent ordering of the rows, use hclust.method = NULL.

Returns:
  @return parameters
}
\examples{
# TODO
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

