% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{populate.radiobuttons}
\alias{populate.radiobuttons}
\title{Description: populate radiobuttons}
\usage{
populate.radiobuttons(tt, title, var.names, var.values, var.init)
}
\arguments{
\item{tt}{TBA}

\item{title}{TBA}

\item{var.names}{TBA}

\item{var.values}{TBA}

\item{var.init}{TBA

Returns:
  @return A list.}
}
\description{
Arguments:
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
