% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune16S.R
\name{prune16S}
\alias{prune16S}
\title{Description: filter 16S data}
\usage{
prune16S(full16S, pmTm.margin = 2.5, complement = 1, mismatch = 0,
  rmoligos = NULL, remove.nonspecific.oligos = FALSE)
}
\description{
Arguments:
  @param full16S full16S
  @param pmTm.margin pmTm margin
  @param complement logical
  @param mismatch logical
  @param rmoligos oligos to exclude
  @param remove.nonspecific.oligos Logical. Remove oligos with multiple targets.
}
\details{
Returns:
  @return filtered 16S data
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

