% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ReadParameters}
\alias{ReadParameters}
\title{Description: Define parameters in select box}
\usage{
ReadParameters(con, which.projects = NULL, all.samples = TRUE,
  chip = NULL)
}
\description{
Arguments:
  @param con Output from dbConnect(dbDriver("MySQL"), username = dbuser, password = dbpwd, dbname = 'PhyloArray')
  @param which.projects Optionally list which projects to take. All samples returned.
  @param all.samples Return all samples from the selected projects: TRUE/FALSE
  @param chip chip
}
\details{
Returns:
  @return list with defined parameters
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

