% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadPhylogeny.R
\name{ReadPhylogeny}
\alias{ReadPhylogeny}
\title{Read phylogeny}
\usage{
ReadPhylogeny(phylogeny, rm.phylotypes, remove.nonspecific.oligos,
  dbuser = dbuser, dbpwd = dbpwd, dbname = dbname, host = host,
  port = port, verbose = verbose, chip)
}
\description{
Arguments:
  @param phylogeny phylogeny
  @param rm.phylotypes rm.phylotypes
  @param remove.nonspecific.oligos remove.nonspecific.oligos
  @param dbuser dbuser		     
  @param dbpwd dbpwd			     
  @param dbname dbname		     
  @param host host		     
  @param port port
  @param verbose verbose
  @param chip chip
}
\details{
Returns:                                        
  @return phylogeny
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

