% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{run.profiling.script}
\alias{run.profiling.script}
\title{run.profiling.script}
\usage{
run.profiling.script(dbuser, dbpwd, dbname, verbose = TRUE, host = NULL,
  port = NULL, summarization.methods = c("frpa", "sum"),
  which.projects = NULL, probe.parameters = NULL)
}
\description{
Description: Profiling main script
}
\details{
Arguments:
  @param dbuser MySQL username
  @param dbpwd  MySQL password
  @param dbname MySQL database name (HITChip: "Phyloarray"; MITChip: "Phyloarray_MIT";
                               PITChip old: "Phyloarray_PIT"; PITChip new: "pitchipdb")
  @param verbose verbose
  @param host host; needed with FTP connections
  @param port port; needed with FTP connections
  @param summarization.methods List summarization methods to be included in output. For HITChip frpa always used; for other chips, rpa always used. Other options: sum, ave
  @param which.projects Optionally specify the projects to extract. All samples from these projects will be included.
  @param probe.parameters probe.parameters

Returns:
  @return Profiling parameters. Also writes output to the user-specified directory.
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

