% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ScaleProfile}
\alias{ScaleProfile}
\title{Description: Between-arrays normalization}
\usage{
ScaleProfile(dat, method = "minmax", bg.adjust = NULL,
  minmax.quantiles = c(0.005, 0.995), minmax.points = NULL)
}
\description{
Arguments:
  @param dat data matrix in original absolute scale
  @param method normalization method
  @param bg.adjust background adjustment 
  @param minmax.quantiles quantiles for minmax
  @param minmax.points minmax end points
}
\details{
Returns:
  @return Normalized data matrix in absolute scale
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

