% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize.probesets.through.species.R
\name{summarize.probesets.through.species}
\alias{summarize.probesets.through.species}
\title{Description: summarize.probesets.through.species}
\usage{
summarize.probesets.through.species(level, taxonomy, oligo.data, method,
  verbose = FALSE)
}
\arguments{
\item{level}{summarization level}

\item{taxonomy}{oligo - phylotype matching data.frame}

\item{oligo.data}{preprocessed probes x samples data matrix in log10 domain}

\item{method}{summarization method}

\item{verbose}{verbose

Returns:}
}
\value{
List with two elements: summarized.matrix (summarized data matrix in log10 scale) and probe.parameters (only used with rpa, probe-level parameter estimates)
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

