% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.rawdata.R
\name{summarize.rawdata}
\alias{summarize.rawdata}
\title{Description: Compress normalized raw data matrix into final probe-level matrix:
             summarize oligos into probes and hybridisations into samples}
\usage{
summarize.rawdata(fdat.log10, fdat.hybinfo, fdat.oligoinfo)
}
\description{
Arguments:
  @param fdat.log10 normalized raw data matrix oligos x hybridisations in log10 scale
  @param fdat.hybinfo hybridization info table
  @param fdat.oligoinfo oligo info table
}
\details{
Returns:
  @return probes x samples matrix in log10 scale
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

