% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.rpa.R
\name{summarize.rpa}
\alias{summarize.rpa}
\title{Description: Probeset summarization with RPA}
\usage{
summarize.rpa(taxonomy, level, probedata, verbose = TRUE,
  probe.parameters = NULL)
}
\description{
Arguments:
  @param taxonomy oligo - phylotype matching data.frame
  @param level taxonomic level for the summarization. 
  @param probedata preprocessed probes x samples data matrix in absolute domain
  @param verbose print intermediate messages
  @param probe.parameters Optional. If probe.parameters are given,
         the summarization is based on these and model parameters are not
	    estimated. A list. One element for each probeset with the following probe vectors: 
    affinities, variances
Returns:
  @return List with two elements: abundance.table (summarized data matrix in absolute scale) and probe.parameters (RPA probe level parameter estimates)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

