% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.sum.R
\name{summarize.sum}
\alias{summarize.sum}
\title{Probeset summarization with SUM}
\usage{
summarize.sum(taxonomy, level, probedata, verbose = TRUE,
  downweight.ambiguous.probes = TRUE)
}
\description{
Arguments:
  @param taxonomy oligo - phylotype matching data.frame
  @param level taxonomic level for the summarization. 
  @param probedata preprocessed probes x samples data matrix in absolute domain
  @param verbose print intermediate messages
  @param downweight.ambiguous.probes Downweight probes with multiple targets
}
\details{
Returns:
  @return List with two elements: abundance.table (summarized data matrix in absolute scale) and probe.parameters used in the calculations
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

