% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{WriteChipData}
\alias{WriteChipData}
\title{Description: Writed data into the output directory}
\usage{
WriteChipData(finaldata, output.dir, tax.table, tax.table.full, meta,
  verbose = TRUE)
}
\description{
Arguments:
  @param finaldata preprocessed data matrices in absolute scale (from the chipdata function)
  @param output.dir output directory
  @param tax.table tax.table used in summarization
  @param tax.table.full tax.table.full unfiltered phylogenyinfo
  @param meta sample metadata samples x features
  @param verbose verbose
}
\details{
Returns:
  @return Preprocessed data in absolute scale, tax.table, and parameters
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

