% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{WriteLog}
\alias{WriteLog}
\title{Description: Write log file}
\usage{
WriteLog(naHybs, params)
}
\description{
Arguments:
  @param naHybs hybridisation that were removed due to NAs  
  @param params parameters
}
\details{
Returns:
  @return List of scaling methods
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

