% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{WriteMatrix}
\alias{WriteMatrix}
\title{Description: Write matrix in tab file}
\usage{
WriteMatrix(dat, filename, verbose = FALSE)
}
\description{
Arguments:
  @param dat data matrix
  @param filename output file
  @param verbose verbose
Returns:
  @return output file location
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

