% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{add.heatmap}
\alias{add.heatmap}
\title{add.heatmap}
\usage{
add.heatmap(
  dat,
  output.dir,
  output.file = NULL,
  taxonomy,
  ppcm = 150,
  hclust.method = "complete",
  palette = "white/black",
  level = "L1",
  metric = "pearson",
  figureratio = 10,
  fontsize = 40,
  tree.display = TRUE
)
}
\arguments{
\item{dat}{oligoprofile data in original (non-log) domain}

\item{output.dir}{output data directory}

\item{output.file}{output file name}

\item{taxonomy}{oligo-phylotype mappings}

\item{ppcm}{figure size}

\item{hclust.method}{hierarchical clustering method}

\item{palette}{color palette ("white/black" / "white/blue" / "black/yellow/white")}

\item{level}{taxonomic level to show}

\item{metric}{clustering metric}

\item{figureratio}{figure ratio}

\item{fontsize}{font size}

\item{tree.display}{tree.display}
}
\value{
Plotting parameters
}
\description{
Add oligprofile heatmap into output directory.
}
\examples{
# data(peerj32); hc <- add.heatmap(peerj32$microbes[, 1:4])
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
