% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{choose.projects}
\alias{choose.projects}
\title{Description: Select projects to analyze}
\usage{
choose.projects(con, multi = TRUE, condition = NULL)
}
\arguments{
\item{con}{valid MySQL connection}

\item{multi}{enable selection of multiple options}

\item{condition}{TBA
Returns:
  @return vector of project names}
}
\description{
Arguments:
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
