% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{estimate.min.threshold}
\alias{estimate.min.threshold}
\title{Description: determine threshold for bg correction}
\usage{
estimate.min.threshold(dat)
}
\arguments{
\item{dat}{data matrix (in approximately normal scale ie. logged)

Returns:}
}
\value{
threshold value
}
\description{
Arguments:
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
