\name{estimate.min.threshold}
\alias{estimate.min.threshold}
\title{Description: determine threshold for bg correction}
\usage{
  estimate.min.threshold(dat)

  estimate.min.threshold(dat)
}
\arguments{
  \item{dat}{data matrix (in approximately normal scale ie.
  logged)

  Returns:}

  \item{dat}{data matrix (in approximately normal scale ie.
  logged)

  Returns:}
}
\value{
  threshold value

  threshold value
}
\description{
  Arguments:

  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}

  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")

  See citation("microbiome")
}
\keyword{utilities}

