% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get.doligo2}
\alias{get.doligo2}
\title{Description: Calculate d.oligo2}
\usage{
get.doligo2(featuretab, d.scaled, oligo.ids)
}
\arguments{
\item{featuretab}{featuretab}

\item{d.scaled}{d.scaled}

\item{oligo.ids}{oligo.ids
Returns:
  @return d.oligo2}
}
\description{
Arguments:
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
