% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{get.phylogeny.info}
\alias{get.phylogeny.info}
\title{get.phylogeny.info}
\usage{
get.phylogeny.info(phylogeny = "16S", dbuser, dbpwd, dbname, verbose = TRUE,
  chip = "HITChip", host = NULL, port = NULL, rmoligos = NULL)
}
\description{
Description: Get phylogeny
}
\details{
Arguments:
  @param phylogeny phylogeny (default: 16S)
  @param dbuser MySQL user
  @param dbpwd MySQL password
  @param dbname MySqL database name
  @param verbose verbose
  @param chip chip type
  @param host host; needed with FTP connections
  @param port port; needed with FTP connections
  @param rmoligos oligos to exclude

Returns:
  @return phylogeny.info
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

