% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HITChipDB-package.R
\docType{package}
\name{HITChipDB-package}
\alias{HITChipDB}
\alias{HITChipDB-package}
\title{R package for HITChipDB studies}
\description{
Brief summary of the HITChipDB package
}
\details{
\tabular{ll}{
Package: \tab HITChipDB\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2012\cr
License: \tab FreeBSD\cr
LazyLoad: \tab yes\cr
}

R package for HITChipDB studies
}
\examples{
library(HITChipDB)
}
\author{
Leo Lahti and Jarkko Salojarvi (UH/WUR). Contact: \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("HITChipDB")
\url{http://microbiome.github.com}
}
\keyword{package}

