% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{list.mysql.projects}
\alias{list.mysql.projects}
\title{list.mysql.projects}
\usage{
list.mysql.projects(dbuser, dbpwd, dbname, host = NULL, port = NULL)
}
\arguments{
\item{dbuser}{MySQL user}

\item{dbpwd}{MySQL password}

\item{dbname}{MySqL database name}

\item{host}{host; needed with FTP connections}

\item{port}{port; needed with FTP connections
Returns:
  @return project names vector}
}
\description{
Description: List projects in MySQL database
}
\details{
Arguments:
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
