% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{oligo.bg.correction}
\alias{oligo.bg.correction}
\title{Description: Calculate species summaries and possibly update d.oligo2}
\usage{
oligo.bg.correction(d.oligo2, bgc.method)

oligo.bg.correction(d.oligo2, bgc.method)
}
\arguments{
\item{d.oligo2}{d.oligo2}

\item{bgc.method}{background correction method
Returns:}

\item{d.oligo2}{d.oligo2}

\item{bgc.method}{background correction method
Returns:}
}
\value{
Background-corrected data matrix

Background-corrected data matrix
}
\description{
Arguments:

Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}

Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")

See citation("microbiome")
}
\keyword{internal}
\keyword{utilities}

