% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{phylotype.rm.list}
\alias{phylotype.rm.list}
\title{Default list of removed phylotypes and oligos}
\usage{
phylotype.rm.list(chip)
}
\arguments{
\item{chip}{Chip name (HIT/MIT/PIT/Chick)Chip}
}
\value{
List of removed oligos and phylotypes
}
\description{
Default list of removed phylotypes and oligos
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

