% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{populate.radiobuttons}
\alias{populate.radiobuttons}
\title{Description: populate radiobuttons}
\usage{
populate.radiobuttons(tt, title, var.names, var.values, var.init)
}
\description{
Arguments:
  @param tt TBA
  @param title TBA
  @param var.names TBA 
  @param var.values TBA
  @param var.init TBA
}
\details{
Returns:
  @return A list.
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

