\name{preprocess.chipdata}
\alias{preprocess.chipdata}
\title{Description: Profiling preprocessing script}
\usage{
  preprocess.chipdata(dbuser, dbpwd, dbname,
    verbose = TRUE, host = NULL, port = NULL,
    use.precalculated.phylogeny = TRUE)
}
\arguments{
  \item{dbuser}{MySQL username}

  \item{dbpwd}{MySQL password}

  \item{dbname}{MySQL database name}

  \item{verbose}{monitor processing through intermediate
  messages}

  \item{host}{host; needed with FTP connections}

  \item{port}{port; needed with FTP connections}

  \item{use.precalculated.phylogeny}{use precalculated
  phylogeny?

  Returns:}
}
\value{
  Preprocessed data and parameters
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

