% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.chipdata.R
\name{preprocess.chipdata}
\alias{preprocess.chipdata}
\title{Description: Profiling preprocessing script}
\usage{
preprocess.chipdata(dbuser, dbpwd, dbname, verbose = TRUE, host = NULL,
  port = NULL, use.precalculated.phylogeny = NULL,
  summarization.methods = c("frpa", "sum"), which.projects = NULL,
  all.samples = TRUE, save.dir = ".", use.default.parameters = FALSE)
}
\arguments{
\item{dbuser}{MySQL username}

\item{dbpwd}{MySQL password}

\item{dbname}{MySQL database name}

\item{verbose}{monitor processing through intermediate messages}

\item{host}{host; needed with FTP connections}

\item{port}{port; needed with FTP connections}

\item{use.precalculated.phylogeny}{use precalculated phylogeny?}

\item{summarization.methods}{List summarization methods to be included in output. With HITChip frpa always used; with other chips rpa always used. Other options: "sum", "ave"}

\item{which.projects}{Optionally specify the projects to extract. All samples from these projects will be included.}

\item{all.samples}{Use all samples from the selected project by default? TRUE / FALSE}

\item{save.dir}{Output data folder}

\item{use.default.parameters}{use.default.parameters}
}
\value{
Preprocessed data and parameters
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

