% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadParameters.R
\name{ReadParameters}
\alias{ReadParameters}
\title{Define parameters in select box}
\usage{
ReadParameters(
  con,
  which.projects = NULL,
  all.samples = TRUE,
  chip = NULL,
  save.dir = NULL,
  use.default.parameters = FALSE
)
}
\arguments{
\item{con}{Output from dbConnect(dbDriver("MySQL"), username = dbuser, password = dbpwd, dbname = 'PhyloArray')}

\item{which.projects}{Optionally list which projects to take. All samples returned.}

\item{all.samples}{Return all samples from the selected projects: TRUE/FALSE}

\item{chip}{chip}

\item{save.dir}{save.dir}

\item{use.default.parameters}{use.default.parameters}
}
\value{
list with defined parameters
}
\description{
Define parameters in select box.
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
