% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ReadParameters}
\alias{ReadParameters}
\title{Description: Define parameters in select box}
\usage{
ReadParameters(con, which.projects = NULL, all.samples = TRUE,
  chip = NULL)
}
\arguments{
\item{con}{Output from dbConnect(dbDriver("MySQL"), username = dbuser, password = dbpwd, dbname = 'PhyloArray')}

\item{which.projects}{Optionally list which projects to take. All samples returned.}

\item{all.samples}{Return all samples from the selected projects: TRUE/FALSE}

\item{chip}{chip

Returns:}
}
\value{
list with defined parameters
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

