% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadPhylogeny.R
\name{ReadPhylogeny}
\alias{ReadPhylogeny}
\title{Read phylogeny}
\usage{
ReadPhylogeny(phylogeny, rm.phylotypes, remove.nonspecific.oligos,
  dbuser = dbuser, dbpwd = dbpwd, dbname = dbname, host = host,
  port = port, verbose = verbose, chip)
}
\arguments{
\item{phylogeny}{phylogeny}

\item{rm.phylotypes}{rm.phylotypes}

\item{remove.nonspecific.oligos}{remove.nonspecific.oligos}

\item{dbuser}{dbuser}

\item{dbpwd}{dbpwd}

\item{dbname}{dbname}

\item{host}{host}

\item{port}{port}

\item{verbose}{verbose}

\item{chip}{chip

Returns:                                        
  @return phylogeny}
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

