% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profiling.R
\name{run.profiling.script}
\alias{run.profiling.script}
\title{run.profiling.script}
\usage{
run.profiling.script(dbuser, dbpwd, dbname, verbose = TRUE, host = NULL,
  port = NULL, summarization.methods = c("frpa", "sum"),
  which.projects = NULL, probe.parameters = NULL)
}
\arguments{
\item{dbuser}{MySQL username}

\item{dbpwd}{MySQL password}

\item{dbname}{MySQL database name (HITChip: "Phyloarray"; MITChip: "Phyloarray_MIT";
                             PITChip old: "Phyloarray_PIT"; PITChip new: "pitchipdb")}

\item{verbose}{verbose}

\item{host}{host; needed with FTP connections}

\item{port}{port; needed with FTP connections}

\item{summarization.methods}{List summarization methods to be included in output. For HITChip frpa always used; for other chips, rpa always used. Other options: sum, ave}

\item{which.projects}{Optionally specify the projects to extract. All samples from these projects will be included.}

\item{probe.parameters}{probe.parameters

Returns:}
}
\value{
Profiling parameters. Also writes output to the user-specified directory.
}
\description{
Description: Profiling main script
}
\details{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

