% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarize.probesets.directly}
\alias{summarize.probesets.directly}
\title{Description: summarize.probesets.directly}
\usage{
summarize.probesets.directly(level, phylogeny.info, oligo.data, method,
  verbose = TRUE)
}
\arguments{
\item{level}{summarization level}

\item{phylogeny.info}{oligo - phylotype matching data.frame}

\item{oligo.data}{preprocessed probes x samples data matrix in log10 domain}

\item{method}{summarization method}

\item{verbose}{verbose

Returns:}
}
\value{
List with two elements: summarized.matrix (summarized data matrix in log10 scale) and probe.parameters (only used with rpa, probe-level parameter estimates)
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

