% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarize.probesets.R
\name{summarize.probesets}
\alias{summarize.probesets}
\title{Description: summarize.probesets}
\usage{
summarize.probesets(phylogeny.info, oligo.data, method, level, verbose = TRUE,
  species.matrix = NULL)
}
\arguments{
\item{phylogeny.info}{oligo - phylotype matching data.frame}

\item{oligo.data}{preprocessed probes x samples data matrix in log10 domain}

\item{method}{summarization method}

\item{level}{summarization level}

\item{verbose}{print intermediate messages}

\item{species.matrix}{Optional. Provide pre-calculated species-level summaries to speed up computation.

Returns:}
}
\value{
List with two elements: summarized.matrix (summarized data matrix in log10 scale) and probe.parameters (only used with rpa, probe-level parameter estimates)
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

