% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize.probesets.species.R
\name{summarize.probesets.species}
\alias{summarize.probesets.species}
\title{Description: Probeset summarization with various methods.}
\usage{
summarize.probesets.species(phylogeny.info, oligo.data, method,
  verbose = TRUE, rm.species = c("Victivallis vadensis"),
  probe.parameters = list())
}
\arguments{
\item{phylogeny.info}{oligo - phylotype matching data.frame}

\item{oligo.data}{preprocessed probes x samples data matrix in log10 domain}

\item{method}{summarization method}

\item{verbose}{print intermediate messages}

\item{rm.species}{Species to exclude}

\item{probe.parameters}{Optional. If probe.parameters are given,
         the summarization is based on these and model parameters are not
        estimated. A list. One element for each probeset with the following probe vectors:
        affinities, variances
Returns:}
}
\value{
List with two elements: summarized.matrix (summarized data matrix in log10 scale) and probe.parameters (only used with rpa, probe-level parameter estimates)
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

