% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.rawdata.R
\name{summarize.rawdata}
\alias{summarize.rawdata}
\title{Description: Compress normalized raw data matrix into final probe-level matrix:
             summarize oligos into probes and hybridisations into samples}
\usage{
summarize.rawdata(fdat.log10, fdat.hybinfo, fdat.oligoinfo)
}
\arguments{
\item{fdat.log10}{normalized raw data matrix oligos x hybridisations in log10 scale}

\item{fdat.hybinfo}{hybridization info table}

\item{fdat.oligoinfo}{oligo info table

Returns:
  @return probes x samples matrix in log10 scale}
}
\description{
Arguments:
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
