% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sync.rm.phylotypes}
\alias{sync.rm.phylotypes}
\title{Description: List oligos associated with removed phylotypes from different levels}
\usage{
sync.rm.phylotypes(rm.phylotypes, tax.table)
}
\arguments{
\item{rm.phylotypes}{rm.phylotypes}

\item{tax.table}{tax.table

Returns:
  @return probe name vector}
}
\description{
Arguments:
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
