% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sync.rm.phylotypes}
\alias{sync.rm.phylotypes}
\title{Description: List oligos associated with removed phylotypes from different levels}
\usage{
sync.rm.phylotypes(rm.phylotypes, phylogeny.info)
}
\arguments{
\item{rm.phylotypes}{rm.phylotypes}

\item{phylogeny.info}{phylogeny.info

Returns:}
}
\value{
probe name vector
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

