% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sync.rm.phylotypes}
\alias{sync.rm.phylotypes}
\title{Description: List oligos associated with removed phylotypes from different levels}
\usage{
sync.rm.phylotypes(rm.phylotypes, tax.table)
}
\description{
Arguments:
  @param rm.phylotypes rm.phylotypes
  @param tax.table tax.table
}
\details{
Returns:
  @return probe name vector
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

