% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{threshold.data}
\alias{threshold.data}
\title{Description: determine detection threshold for the data}
\usage{
threshold.data(dat, sd.times = 6)
}
\arguments{
\item{dat}{data}

\item{sd.times}{standard deviation threshold

Returns:
  @return thresholded data matrix}
}
\description{
Arguments:
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
