% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{threshold.data}
\alias{threshold.data}
\title{Description: determine detection threshold for the data}
\usage{
threshold.data(dat, sd.times = 6)
}
\description{
Arguments:
  @param dat data
  @param sd.times standard deviation threshold
}
\details{
Returns:
  @return thresholded data matrix
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

