% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{WriteChipData}
\alias{WriteChipData}
\title{Description: Writed data into the output directory}
\usage{
WriteChipData(finaldata, output.dir, phylogeny.info, phylogeny.info.full,
  verbose = TRUE)
}
\arguments{
\item{finaldata}{preprocessed data matrices in absolute scale (from the chipdata function)}

\item{output.dir}{output directory}

\item{phylogeny.info}{phylogeny.info used in summarization}

\item{phylogeny.info.full}{phylogeny.info.full unfiltered phylogenyinfo}

\item{verbose}{verbose

Returns:}
}
\value{
Preprocessed data in absolute scale, phylogeny.info, and parameters
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

