% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{WriteMatrix}
\alias{WriteMatrix}
\title{Description: Write matrix in tab file}
\usage{
WriteMatrix(dat, filename, verbose = FALSE)
}
\arguments{
\item{dat}{data matrix}

\item{filename}{output file}

\item{verbose}{verbose}
}
\value{
output file location
}
\description{
Description: Write matrix in tab file
}
\references{
See citation("microbiome")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
